+++ core/src/com/ct/game/model/entities/WaterTile.java
@@ -13,7 +13,7 @@ import com.ct.game.view.GameScreen;
 public class WaterTile extends Tile {
 
     @Override
-    public void init(int row, int col) {
+    public void init(float row, float col) {
         super.init(row, col);
 
         add(new PhysicsComponent(
